+++ Editor/SettingsEditor.cs
@@ -1,16 +1,16 @@
-﻿//#define ENABLE_LEGACY_PREFERENCE
+﻿#define ENABLE_HANANOKI_SETTINGS
 
 using UnityEditor;
 using UnityEngine;
-using Hananoki.Extensions;
-using Hananoki.SharedModule;
+using HananokiEditor.Extensions;
+using HananokiEditor.SharedModule;
 using UnityEditorInternal;
 using System.Collections.Generic;
 
-using E = Hananoki.ManifestJsonUtility.SettingsEditor;
-using SS = Hananoki.SharedModule.S;
+using E = HananokiEditor.ManifestJsonUtility.SettingsEditor;
+using SS = HananokiEditor.SharedModule.S;
 
-namespace Hananoki.ManifestJsonUtility {
+namespace HananokiEditor.ManifestJsonUtility {
 	[System.Serializable]
 	public class SettingsEditor {
 
@@ -143,13 +143,12 @@ namespace Hananoki.ManifestJsonUtility {
 
 
 #if ENABLE_HANANOKI_SETTINGS
-	[SettingsClass]
 	public class SettingsEvent {
-		[SettingsMethod]
+		[HananokiSettingsRegister]
 		public static SettingsItem RegisterSettings() {
 			return new SettingsItem() {
 				//mode = 1,
-				displayName = Package.name,
+				displayName = Package.nameNicify,
 				version = Package.version,
 				gui = SettingsEditorWindow.DrawGUI,
 			};